each command gets different response handler
